Disabled choose length when starting timer
For the same reason as disabling the start button. When user clicks
reset both buttons will be enabled again. Also fiddled with some
settings to make things responsive. Chrome devtools gives a different
picture to just resizing the screen so I think I will have to host it
and check using a mobile device myself to be sure of what the circle
looks like on a small screen. Might have to abandon the circle on
smaller devices